+++ persistence/test/setup_test.go
@@ -21,6 +21,7 @@ import (
 	coreTypes "github.com/pokt-network/pocket/shared/core/types"
 	"github.com/pokt-network/pocket/shared/messaging"
 	"github.com/pokt-network/pocket/shared/modules"
+	moduleTypes "github.com/pokt-network/pocket/shared/modules/types"
 	"github.com/stretchr/testify/require"
 	"golang.org/x/exp/slices"
 )
@@ -42,8 +43,8 @@ var (
 	StakeToUpdate        = converters.BigIntToString((&big.Int{}).Add(DefaultStakeBig, DefaultDeltaBig))
 
 	DefaultStakeStatus     = int32(persistence.StakedStatus)
-	DefaultPauseHeight     = int64(-1)
-	DefaultUnstakingHeight = int64(-1)
+	DefaultPauseHeight     = int64(-1) // pauseHeight=-1 means not paused
+	DefaultUnstakingHeight = int64(-1) // pauseHeight=-1 means not unstaking
 
 	OlshanskyURL    = "https://olshansky.info"
 	OlshanskyChains = []string{"OLSH"}
@@ -235,8 +236,8 @@ func fuzzSingleProtocolActor(
 			if originalActor.UnstakingHeight != db.Height { // Not ready to unstake
 				require.Nil(t, unstakingActors)
 			} else {
-				idx := slices.IndexFunc(unstakingActors, func(a modules.IUnstakingActor) bool {
-					return originalActor.Address == hex.EncodeToString(a.GetAddress())
+				idx := slices.IndexFunc(unstakingActors, func(a *moduleTypes.UnstakingActor) bool {
+					return originalActor.Address == a.Address
 				})
 				require.NotEqual(t, idx, -1, fmt.Sprintf("actor that is unstaking was not found %+v", originalActor))
 			}
@@ -285,8 +286,8 @@ func fuzzSingleProtocolActor(
 			newActor, err := getTestActor(db, originalActor.Address)
 			require.NoError(t, err)
 
-			if db.Height > originalActor.PausedHeight { // isPausedAndReadyToUnstake
-				require.Equal(t, newActor.UnstakingHeight, newUnstakingHeight, "setPausedToUnstaking")
+			if db.Height > originalActor.PausedHeight && originalActor.PausedHeight != DefaultPauseHeight { // isPausedAndReadyToUnstake
+				require.Equal(t, newUnstakingHeight, newActor.UnstakingHeight, "setPausedToUnstaking")
 			}
 		case "GetActorOutputAddr":
 			outputAddr, err := db.GetActorOutputAddress(protocolActorSchema, addr, db.Height)